+++ ElvUI_MerathilisUI/Core/General/Changelog.lua
@@ -12,9 +12,7 @@ local DISABLED_FONT_COLOR = DISABLED_FONT_COLOR
 
 local ChangeLogData = {
 	"Changes:",
-		"• FOR 10.0.7",
-		"• Updated chat code to reflect 10.0.7 changes",
-		"• Updated MicroBar to the latest API Changes",
+		"• Chat: Fix a non Retail Chat error.. narf",
 
 		" ",