+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 ### Changes ###
 
-  * Updated chat code to reflect 10.0.7 changes
-  * Updated MicroBar to the latest API Changes
+  * Chat: Fix a non Retail Chat error.. narf